sconv: Improve performance of s24 <-> s16 C implementation by 10%

+++ src/pulsecore/sconv_c.h
@@ -36,8 +36,8 @@ void pa_sconv_s32le_from_s16ne(unsigned n, const int16_t *a, int32_t *b);
 void pa_sconv_s32le_to_float32ne(unsigned n, const int32_t *a, float *b);
 void pa_sconv_s32le_from_float32ne(unsigned n, const float *a, int32_t *b);
 
-void pa_sconv_s24le_to_s16ne(unsigned n, const uint8_t *a, int16_t *b);
-void pa_sconv_s24le_from_s16ne(unsigned n, const int16_t *a, uint8_t *b);
+void pa_sconv_s24le_to_s16ne(unsigned n, const uint8_t *a, uint8_t *b);
+void pa_sconv_s24le_from_s16ne(unsigned n, const uint8_t *a, uint8_t *b);
 void pa_sconv_s24le_to_float32ne(unsigned n, const uint8_t *a, float *b);
 void pa_sconv_s24le_from_float32ne(unsigned n, const float *a, uint8_t *b);
 
@@ -58,8 +58,8 @@ void pa_sconv_s32be_from_s16ne(unsigned n, const int16_t *a, int32_t *b);
 void pa_sconv_s32be_to_float32ne(unsigned n, const int32_t *a, float *b);
 void pa_sconv_s32be_from_float32ne(unsigned n, const float *a, int32_t *b);
 
-void pa_sconv_s24be_to_s16ne(unsigned n, const uint8_t *a, int16_t *b);
-void pa_sconv_s24be_from_s16ne(unsigned n, const int16_t *a, uint8_t *b);
+void pa_sconv_s24be_to_s16ne(unsigned n, const uint8_t *a, uint8_t *b);
+void pa_sconv_s24be_from_s16ne(unsigned n, const uint8_t *a, uint8_t *b);
 void pa_sconv_s24be_to_float32ne(unsigned n, const uint8_t *a, float *b);
 void pa_sconv_s24be_from_float32ne(unsigned n, const float *a, uint8_t *b);
